refactor(web): collapse campaign jobs prop tunnel; move install commands
