+++ apps/web/src/components/features/campaigns/campaign-detail.tsx
@@ -3,20 +3,12 @@
 import type { ReactElement } from "react";
 import { LinearProgress, Stack } from "@mui/material";
 import { useQueryClient } from "@tanstack/react-query";
-import type { Route } from "next";
-import { useRouter } from "next/navigation";
-import { api } from "@/api/client";
-import { apiErrorMessage } from "@/api/error";
 import { useApiQuery } from "@/api/hooks";
 import { campaignQueries } from "@/api/queries";
 import { queryKeys } from "@/api/query-keys";
-import type { CampaignJobDto } from "@/api/types";
 import { OutreachBoard } from "@/components/features/outreach";
 import { campaignChannel } from "@/lib/sse/channels/campaign";
 import { useSseChannel } from "@/lib/sse/client";
-import { useAgent, useAgentAvailable } from "@/providers/agent-provider";
-import { useToast } from "@/providers/notification-provider";
-import { UPWORK_DOMAIN } from "./composer/form-config";
 import { CampaignHeaderCard } from "./detail/header-card";
 import { CampaignJobsPanel } from "./detail/jobs-panel";
 import { CampaignReasonBreakdown } from "./detail/reason-breakdown";
@@ -28,11 +20,7 @@ interface CampaignDetailProps {
 
 export function CampaignDetail(props: CampaignDetailProps): ReactElement {
   const { campaignId } = props;
-  const router = useRouter();
   const queryClient = useQueryClient();
-  const agent = useAgent();
-  const agentAvailable = useAgentAvailable();
-  const toast = useToast();
 
   const detail = useApiQuery(campaignQueries.detail(campaignId));
 
@@ -56,8 +44,6 @@ export function CampaignDetail(props: CampaignDetailProps): ReactElement {
   }
 
   const campaign = detail.data;
-  const isAutoApply = campaign.source === "auto-apply";
-  const isUpwork = campaign.config.board === UPWORK_DOMAIN;
 
   if (campaign.source === "outreach") {
     return (
@@ -73,46 +59,12 @@ export function CampaignDetail(props: CampaignDetailProps): ReactElement {
     );
   }
 
-  // Auto-apply campaigns apply on their own; on other campaigns (e.g. search results) the
-  // user dispatches a job to the single-job apply flow by its URL.
-  const applyJob = (job: CampaignJobDto): void => {
-    if (!isAutoApply) {
-      return;
-    }
-    void agent.injectSkill("apply", job.url);
-  };
-
-  // Upwork recommendations are recommend-only: seed a proposal draft from the
-  // recommendation, then hand off to the upwork-proposal skill to write it.
-  const draftProposal = async (job: CampaignJobDto): Promise<void> => {
-    const res = await api.upwork.proposals.post({
-      jobTitle: job.title,
-      clientName: job.company || null,
-      jobUrl: job.url,
-      jobDescription: job.description ?? "",
-      source: "search",
-      campaignId: job.campaignId,
-      jobKey: job.key,
-    });
-    if (res.error || !res.data) {
-      toast.error(apiErrorMessage(res.error, "Could not create the proposal draft"));
-      return;
-    }
-    void agent.injectSkill("upwork-proposal", String(res.data.id));
-    router.push(`/upwork/${res.data.id}` as Route);
-  };
-
   return (
     <Stack spacing={3}>
       <CampaignHeaderCard campaign={campaign} />
       <CampaignSummaryTiles campaign={campaign} />
       <CampaignReasonBreakdown campaign={campaign} />
-      <CampaignJobsPanel
-        campaign={campaign}
-        onApplyJob={!isUpwork && agentAvailable ? applyJob : undefined}
-        onDraftProposal={isUpwork && agentAvailable ? draftProposal : undefined}
-        showReason={isUpwork}
-      />
+      <CampaignJobsPanel campaign={campaign} />
     </Stack>
   );
 }
