+++ apps/web/src/components/features/install/install-commands.ts
